requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
certifi==2024.12.14
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.18.1
jedi==0.19.2
jiter==0.8.2
matplotlib-inline==0.1.7
openai==1.58.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dotenv==1.0.1
sniffio==1.3.1
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
